version 18// Customer Search by Catgeory // paging // Sorting // Product Images "more than one image base 64"

+++ GradProJM/src/main/java/com/example/GradProJM/Services/ProductShopService.java
@@ -2,7 +2,6 @@ package com.example.GradProJM.Services;
 
 import com.example.GradProJM.Model.*;
 import com.example.GradProJM.Repos.*;
-import org.aspectj.weaver.ast.Or;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
@@ -13,7 +12,6 @@ import org.springframework.stereotype.Service;
 import java.text.DecimalFormat;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
@@ -33,19 +31,19 @@ public class ProductShopService {
                               OrderRepository ordRepo,
                               customerFeedbackRepository custfeedbkRepo,
                               CustomerRepository custRepo) {
-        this.prdshpRepo=prdshpRepo;
-        this.prdRepo=prdRepo;
-        this.shpRepo=shpRepo;
-        this.ordRepo=ordRepo;
-        this.custfeedbkRepo=custfeedbkRepo;
-        this.custRepo=custRepo;
+        this.prdshpRepo = prdshpRepo;
+        this.prdRepo = prdRepo;
+        this.shpRepo = shpRepo;
+        this.ordRepo = ordRepo;
+        this.custfeedbkRepo = custfeedbkRepo;
+        this.custRepo = custRepo;
     }
 
     public Boolean AddAnExistingProducttoaShop(Shop_Products shopProducts) {
-        Optional<product> product=prdRepo.findById(shopProducts.getProduct().getProductId());
-        if(product.isPresent()){
-            Optional<ShopOwner> shop=shpRepo.findById(shopProducts.getShop().getShopID());
-            if(shop.isPresent()){
+        Optional<product> product = prdRepo.findById(shopProducts.getProduct().getProductId());
+        if (product.isPresent()) {
+            Optional<ShopOwner> shop = shpRepo.findById(shopProducts.getShop().getShopID());
+            if (shop.isPresent()) {
                 prdshpRepo.save(shopProducts);
                 return true;
             }
@@ -55,16 +53,16 @@ public class ProductShopService {
     }
 
     public Boolean AddNewProducttoaShop(Shop_Products shopProducts) {
-        Optional<product> checkprod=prdRepo.findByproductBarcode(shopProducts.getProduct().getProductBarcode());
-        int flag=0;
-        List<Shop_Products> chk=prdshpRepo.findAll();
-        for(Shop_Products sh: chk){
-            if(sh.getProduct().getProductBarcode().equals(shopProducts.getProduct().getProductBarcode()) &&
-                    sh.getShop().getShopID()==shopProducts.getShop().getShopID()){
-                flag=1;
+        Optional<product> checkprod = prdRepo.findByproductBarcode(shopProducts.getProduct().getProductBarcode());
+        int flag = 0;
+        List<Shop_Products> chk = prdshpRepo.findAll();
+        for (Shop_Products sh : chk) {
+            if (sh.getProduct().getProductBarcode().equals(shopProducts.getProduct().getProductBarcode()) &&
+                    sh.getShop().getShopID() == shopProducts.getShop().getShopID()) {
+                flag = 1;
             }
         }
-        if(flag==0) {
+        if (flag == 0) {
             if (checkprod.isPresent()) {
                 return null;
             } else {
@@ -78,35 +76,35 @@ public class ProductShopService {
     }
 
     public Boolean AddAnExistingProductByBarcodetoaShop(Shop_Products shopProducts) {
-        Optional<product> product=prdRepo.findByproductBarcode(shopProducts.getProduct().getProductBarcode());
+        Optional<product> product = prdRepo.findByproductBarcode(shopProducts.getProduct().getProductBarcode());
         Optional<Shop_Products> shsh = prdshpRepo.findShop_ProductsByShop_ShopNameAndProduct_ProductBarcode(shopProducts.getShop().getShopName(),
                 shopProducts.getProduct().getProductBarcode());
-        int flag=0;
-        List<Shop_Products> chk=prdshpRepo.findAll();
-        for(Shop_Products sh: chk){
-            if(sh.getProduct().getProductBarcode().equals(shopProducts.getProduct().getProductBarcode()) &&
-            sh.getShop().getShopID()==shopProducts.getShop().getShopID()){
-                flag=1;
+        int flag = 0;
+        List<Shop_Products> chk = prdshpRepo.findAll();
+        for (Shop_Products sh : chk) {
+            if (sh.getProduct().getProductBarcode().equals(shopProducts.getProduct().getProductBarcode()) &&
+                    sh.getShop().getShopID() == shopProducts.getShop().getShopID()) {
+                flag = 1;
             }
         }
-        if(flag==0) {
-        if(product.isPresent()){
-            Optional<ShopOwner> shop=shpRepo.findById(shopProducts.getShop().getShopID());
-            if(shop.isPresent()){
-                System.out.println(product.get());
-                Shop_Products shpprdct= new Shop_Products(shopProducts.getShop(),product.get(),shopProducts.getQuantity());
-                shpprdct.setDeleted(false);
-                shpprdct.setProductPrice(shopProducts.getProductPrice());
-                shpprdct.setProductDescription(shopProducts.getProductDescription());
-                shpprdct.setProductPublishDate(String.valueOf(LocalDateTime.now()));
-                prdshpRepo.save(shpprdct);
-                return true;
+        if (flag == 0) {
+            if (product.isPresent()) {
+                Optional<ShopOwner> shop = shpRepo.findById(shopProducts.getShop().getShopID());
+                if (shop.isPresent()) {
+                    System.out.println(product.get());
+                    Shop_Products shpprdct = new Shop_Products(shopProducts.getShop(), product.get(), shopProducts.getQuantity());
+                    shpprdct.setDeleted(false);
+                    shpprdct.setProductPrice(shopProducts.getProductPrice());
+                    shpprdct.setProductDescription(shopProducts.getProductDescription());
+                    shpprdct.setProductPublishDate(String.valueOf(LocalDateTime.now()));
+                    prdshpRepo.save(shpprdct);
+                    return true;
+                }
+                return null;
             }
-            return null;
-        }
 
         }
-        if(flag ==1 && shsh.get().isDeleted()){
+        if (flag == 1 && shsh.get().isDeleted()) {
             shsh.get().setDeleted(false);
             shsh.get().setProductPrice(shopProducts.getProductPrice());
             shsh.get().setQuantity(shopProducts.getQuantity());
@@ -123,14 +121,14 @@ public class ProductShopService {
     }
 
     public Boolean DeleteAProductFromAShop(int shopID, int prodID) {
-        Optional<ShopOwner> shop=shpRepo.findById(shopID);
+        Optional<ShopOwner> shop = shpRepo.findById(shopID);
         List<Shop_Products> shopProducts = prdshpRepo.findAll();
-        if(shop.isPresent()){
-            Optional<product> prod=prdRepo.findById(prodID);
-            if(prod.isPresent()){
-                for(Shop_Products ch : shopProducts){
-                    if(ch.getShop().getShopID()==shopID &&
-                    ch.getProduct().getProductId()==prodID){
+        if (shop.isPresent()) {
+            Optional<product> prod = prdRepo.findById(prodID);
+            if (prod.isPresent()) {
+                for (Shop_Products ch : shopProducts) {
+                    if (ch.getShop().getShopID() == shopID &&
+                            ch.getProduct().getProductId() == prodID) {
                         prdshpRepo.delete(ch);
                         return true;
                     }
@@ -143,19 +141,20 @@ public class ProductShopService {
     }
 
     public Boolean deleteAProductByBarcodeFromAShop(String shopName, String prodBarcode) {
-        Optional <ShopOwner> shop = shpRepo.findShopOwnerByShopName(shopName);
-        if(shop.isPresent()){
+        Optional<ShopOwner> shop = shpRepo.findShopOwnerByShopName(shopName);
+        if (shop.isPresent()) {
             List<Shop_Products> products = shop.get().getShopProducts();
             Optional<Shop_Products> prod = prdshpRepo.findShop_ProductsByShop_ShopNameAndProduct_ProductBarcode(
-                    shopName,prodBarcode
+                    shopName, prodBarcode
             );
-            if(prod.isPresent()){
-            products.remove(prod.get());
-            shop.get().setShopProducts(products);
-            shpRepo.save(shop.get());
-            prod.get().setDeleted(true);
-            prdshpRepo.save(prod.get());
-            return true;}
+            if (prod.isPresent()) {
+                products.remove(prod.get());
+                shop.get().setShopProducts(products);
+                shpRepo.save(shop.get());
+                prod.get().setDeleted(true);
+                prdshpRepo.save(prod.get());
+                return true;
+            }
             return false;
         }
         return false;
@@ -185,7 +184,6 @@ public class ProductShopService {
 //    }
 
 
-
     public Order MakeNewOrder(Order order) {
         Order ord = new Order();
         ord.setCustomer(order.getCustomer());
@@ -209,12 +207,12 @@ public class ProductShopService {
 //            }
         return null;
 
-        }
+    }
 
     public Shop_Products rateAProduct(customerRates custRate) {
         Optional<ShopOwner> shop = shpRepo.findById(custRate.getProducts().getShop().getShopID());
         Optional<product> product = prdRepo.findById(custRate.getProducts().getProduct().getProductId());
-        if(shop.isPresent() && product.isPresent()) {
+        if (shop.isPresent() && product.isPresent()) {
             Optional<Shop_Products> shopProduct = prdshpRepo.findShop_ProductsByShop_ShopNameAndProduct_ProductBarcode(shop.get().getShopName(), product.get().getProductBarcode());
             if (shopProduct.isPresent()) {
                 Optional<product> prod = prdRepo.findByproductBarcode(product.get().getProductBarcode());
@@ -229,7 +227,7 @@ public class ProductShopService {
                 shopProduct.get().setRate(rt);
                 shopProduct.get().setNumOfRates(noRates);
                 shopProduct.get().setRate(rt);
-                Optional <Customer> customer = custRepo.findById(custRate.getCustomer().getCustID());
+                Optional<Customer> customer = custRepo.findById(custRate.getCustomer().getCustID());
                 custRate.setProducts(shopProduct.get());
                 custRate.setCustomer(customer.get());
                 custfeedbkRepo.save(custRate);
@@ -238,27 +236,28 @@ public class ProductShopService {
                 return shopProduct.get();
                 //todo: save the customer..
                 //todo: send an confirmation link to the email ...
-                
+
 
             }
         }
-            return null;
+        return null;
 
     }
 
     public Shop_Products rateAProduct(int custID, String shopName, String prodBarcode, customerRates custRate) {
         Optional<Shop_Products> shopProduct = prdshpRepo.findShop_ProductsByShop_ShopNameAndProduct_ProductBarcode(shopName, prodBarcode);
-        Optional <Customer> customer = custRepo.findById(custID);
+        Optional<Customer> customer = custRepo.findById(custID);
         custRate.setCustomer(customer.get());
         custRate.setProducts(shopProduct.get());
         return rateAProduct(custRate);
     }
 
     public List<Shop_Products> view(int shopID, int pageNum) {
-        Optional<List<Shop_Products>> products = prdshpRepo.findShop_ProductsByShop_ShopIDAndDeletedFalse(shopID, PageRequest.of(pageNum,2));
+        Optional<List<Shop_Products>> products = prdshpRepo.findShop_ProductsByShop_ShopIDAndDeletedFalse(shopID, PageRequest.of(pageNum, 2));
         return products.get();
     }
-    public int getShopProductsPagesNum(int shopID){
+
+    public int getShopProductsPagesNum(int shopID) {
         long totalProducts = prdshpRepo.countByShopShopIDAndDeletedFalse(shopID);
         return (int) Math.ceil((double) totalProducts / 2);
         //todo: change 2 to 10
@@ -266,15 +265,15 @@ public class ProductShopService {
 
 
     public List<Shop_Products> getProductsbyCategory(String category, int shopID, int pageNum) {
-        Optional<List<Shop_Products>> products = prdshpRepo.findShop_ProductsByShop_ShopIDAndProductProductCategoryAndDeletedFalse(shopID, category, PageRequest.of(pageNum,2));
+        Optional<List<Shop_Products>> products = prdshpRepo.findShop_ProductsByShop_ShopIDAndProductProductCategoryAndDeletedFalse(shopID, category, PageRequest.of(pageNum, 2));
         return products.get();
     }
-    public int getCategoriesPagesNum(String cat, int shopID){
+
+    public int getCategoriesPagesNum(String cat, int shopID) {
         long totalProducts = prdshpRepo.countByShopShopIDAndProductProductCategoryAndDeletedFalse(shopID, cat);
         return (int) Math.ceil((double) totalProducts / 2);
         //todo: change 2 to 10
     }
-
 
 
     public Page<Shop_Products> SortASC(int shopID, int pageNum, String field) {
@@ -287,9 +286,9 @@ public class ProductShopService {
         return prdshpRepo.findShop_ProductsByShop_ShopIDAndDeletedFalse(shopID, pageable);
     }
 
-    public Shop_Products updateProduct(Shop_Products product,String shopName, String prodBarcode) {
+    public Shop_Products updateProduct(Shop_Products product, String shopName, String prodBarcode) {
         Optional<Shop_Products> shopProduct = prdshpRepo.findShop_ProductsByShop_ShopNameAndProduct_ProductBarcode(shopName, prodBarcode);
-        if(shopProduct.isPresent()){
+        if (shopProduct.isPresent()) {
             shopProduct.get().setProductPrice(product.getProductPrice());
             shopProduct.get().setQuantity(product.getQuantity());
             prdshpRepo.save(shopProduct.get());
@@ -314,7 +313,7 @@ public class ProductShopService {
     }
 
     public List<Shop_Products> SearchProducts(String prodName, int pageNum) {
-        Optional<List<Shop_Products>> products = prdshpRepo.findShop_ProductsByProductProductNameStartingWithAndDeletedFalse(prodName, PageRequest.of(pageNum,2));
+        Optional<List<Shop_Products>> products = prdshpRepo.findShop_ProductsByProductProductNameStartingWithAndDeletedFalse(prodName, PageRequest.of(pageNum, 2));
         return products.get();
     }
 
@@ -322,4 +321,41 @@ public class ProductShopService {
         long totalProducts = prdshpRepo.countByProductProductNameStartingWithAndDeletedFalse(prodName);
         return (int) Math.ceil((double) totalProducts / 2);
     }
+
+    public Page<Shop_Products> sortCategoryPageASC(String category, int shopID, String field, int pageNum) {
+        Pageable pageable = PageRequest.of(pageNum, 2, Sort.by(Sort.Direction.DESC, field));
+        return prdshpRepo.findShop_ProductsByShop_ShopIDAndProductProductCategoryAndDeletedFalse(shopID, category, pageable);
+    }
+
+    public Page<Shop_Products> sortCategoryPageDESC(String category, int shopID, String field, int pageNum) {
+        Pageable pageable = PageRequest.of(pageNum, 2, Sort.by(Sort.Direction.ASC, field));
+        return prdshpRepo.findShop_ProductsByShop_ShopIDAndProductProductCategoryAndDeletedFalse(shopID, category, pageable);
+    }
+
+    public List<Shop_Products> SearchProductsByCatt(String category, int pageNum) {
+        Optional<List<Shop_Products>> products = prdshpRepo.findShop_ProductsByProductProductCategoryAndDeletedFalse(category, PageRequest.of(pageNum, 2));
+        return products.get();
+    }
+
+    public int getSearchByCattPagesNum(String cat) {
+        long totalProducts = prdshpRepo.countByProductProductCategoryAndDeletedFalse(cat);
+        return (int) Math.ceil((double) totalProducts / 2);
+        //todo: change 2 to 10
+    }
+
+    public Page<Shop_Products> sortCatgeoryASC(String category, int pageNum, String field) {
+        Pageable pageable = PageRequest.of(pageNum, 2, Sort.by(Sort.Direction.ASC, field));
+        return prdshpRepo.findShop_ProductsByProductProductCategoryAndDeletedFalse(category, pageable);
+
+    }
+
+    public Page<Shop_Products> sortCatgeoryDESC(String category, int pageNum, String field) {
+        Pageable pageable = PageRequest.of(pageNum, 2, Sort.by(Sort.Direction.DESC, field));
+        return prdshpRepo.findShop_ProductsByProductProductCategoryAndDeletedFalse(category, pageable);
+    }
+
+
+
+
+
 }
